requirements.txt
# Data processing & ETL
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Apache Spark (batch + streaming)
pyspark>=3.5.0

# Snowflake
snowflake-connector-python>=3.8.0
snowflake-sqlalchemy>=1.11.0

# Database & SQL
sqlalchemy>=2.0.0
sqlparse>=0.4.0

# Dashboard & visualization
streamlit>=1.28.0
plotly>=5.18.0
altair>=5.0.0

# Config & env
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Power BI data export
openpyxl>=3.1.0

# Utilities
schedule>=1.2.0
requests>=2.31.0
tenacity>=8.2.0
